further changes

+++ scripts/exrtma3d_updatevars.ksh
@@ -24,21 +24,10 @@ if [ "${envir}" == "esrl" ]; then
 fi
 
 # make sure executable exists
-if [ ! -f ${EXECrtma3d}/${exefile_name_updatevars_wrf} ]; then
-  ${ECHO} "ERROR: executable '${EXECrtma3d}/${exefile_name_updatevars_wrf}' does not exist!"
+if [ ! -f ${EXECrtma3d}/${exefile_name_updatevars} ]; then
+  ${ECHO} "ERROR: executable '${EXECrtma3d}/${exefile_name_updatevars}' does not exist!"
   exit 1
 fi
-if [ ! -f ${EXECrtma3d}/${exefile_name_updatevars_ncfields} ]; then
-  ${ECHO} "ERROR: executable '${EXECrtma3d}/${exefile_name_updatevars_ncfields}' does not exist!"
-  exit 1
-fi
-if [ ! -f ${EXECrtma3d}/${exefile_name_updatevars_ndown} ]; then
-  ${ECHO} "ERROR: executable '${EXECrtma3d}/${exefile_name_updatevars_ndown}' does not exist!"
-  exit 1
-fi
-
-
-
 
 # Check to make sure required directory defined and existed
 check_if_defined "FCST_LENGTH" "GSIRUN_DIR" "FIXwrf" "PDY" "cyc" "subcyc"
@@ -228,24 +217,28 @@ if [ ! -e "wrfout_d01_${time_str}" ]; then
   exit 1
 fi 
 
-${LN} -s wrfout_d01_${time_str} wrfout_d01
-${CP_LN} ${EXECrtma3d}/${exefile_name_update_ncfields} .
-
 # Output successful so write status to log
 ${ECHO} "Assemble Reflectivity fields back into wrf_inout"
+if [ "${envir}" == "esrl" ]; then #Jet
+  #${NCKS} -A -H -v REFL_10CM,COMPOSITE_REFL_10CM,REFL_10CM_1KM,REFL_10CM_4KM,U10,V10 wrfout_d01_${time_str} ${GSIRUN_DIR}/wrf_inout
+  ${NCKS} -A -H -v REFL_10CM,COMPOSITE_REFL_10CM,REFL_10CM_1KM,REFL_10CM_4KM wrfout_d01_${time_str} ${GSIRUN_DIR}/wrf_inout
+else
+  if [ ! -f ${EXECrtma3d}/${exefile_name_updatevars_ncfields} ]; then
+    ${ECHO} "ERROR: executable '${EXECrtma3d}/${exefile_name_updatevars_ncfields}' does not exist!"
+    exit 1
+  fi
+  ${CP_LN} ${EXECrtma3d}/${exefile_name_update_ncfields} .
+  ${LN} -s wrfout_d01_${time_str} wrfout_d01
+  ${exefile_name_update_ncfields} wrfout_d01 wrf_inout 
+  export err=$?; err_chk
 
-#${NCKS} -A -H -v REFL_10CM,COMPOSITE_REFL_10CM,REFL_10CM_1KM,REFL_10CM_4KM,U10,V10 wrfout_d01_${time_str} ${COMOUTgsi_rtma3d}/${ANLrtma3d_FNAME} 
-#${NCKS} -A -H -v REFL_10CM,COMPOSITE_REFL_10CM,REFL_10CM_1KM,REFL_10CM_4KM wrfout_d01_${time_str} ${COMOUTgsi_rtma3d}/${ANLrtma3d_FNAME} 
+  if [ -f ${COMOUTgsi_rtma3d}/${ANLrtma3d_FNAME} ]; then
+    ${ECHO} "Erasing the GSI generated analysis file to be replaced by modified analysis."
+    ${RM} ${COMOUTgsi_rtma3d}/${ANLrtma3d_FNAME}
+  fi
 
-${exefile_name_update_ncfields} wrfout_d01 wrf_inout 
-export err=$?; err_chk
-
-if [ -f ${COMOUTgsi_rtma3d}/${ANLrtma3d_FNAME} ]; then
-  ${ECHO} "Erasing the GSI generated analysis file to be replaced by modified analysis."
-  ${RM} ${COMOUTgsi_rtma3d}/${ANLrtma3d_FNAME}
+  ${CP_LN} -p wrf_inout ${COMOUTgsi_rtma3d}/${ANLrtma3d_FNAME}
 fi
-
-${CP_LN} -p wrf_inout ${COMOUTgsi_rtma3d}/${ANLrtma3d_FNAME}
 
 ${ECHO} "update_vars.ksh completed successfully at `${DATE}`"
